Rewrote heuristic function to improve playing
Old heuristic function looked at how many 3 in a rows we could form where there was a possibility to play a 4th chip. New heuristic takes into account 2 in a rows where you can play chips in the 2 empty slots, 3 in a rows where you can play a chip in the empty slot, and also OPPONENT 3 in a rows which we penalize for.